# reference/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Prebet UPSI Ride-Hailing API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ratings:
    post:
      operationId: submitRating
      tags: [ratings]
      summary: Submit a rating for a completed ride
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RatingInput"
      responses:
        "201":
          description: Rating submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rating"
  /ratings/driver/{driverId}:
    get:
      operationId: getDriverRatings
      tags: [ratings]
      summary: Get ratings for a driver
      parameters:
        - name: driverId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ratings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rating"

  # Dashboard / wow endpoints
...
components:
  schemas:
...
    Rating:
      type: object
      required: [id, bookingId, driverId, studentId, rating, createdAt]
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        driverId:
          type: integer
        studentId:
          type: integer
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        studentName:
          type: string
        createdAt:
          type: string
    RatingInput:
      type: object
      required: [bookingId, driverId, rating]
      properties:
        bookingId:
          type: integer
        driverId:
          type: integer
        rating:
          type: integer
        comment:
          type: string

    # Dashboard
...