# reference/lib/api-zod/src/generated/types/ratingInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Prebet UPSI Ride-Hailing API
 * OpenAPI spec version: 0.1.0
 */

export interface RatingInput {
  bookingId: number;
  driverId: number;
  rating: number;
  comment?: string;
}